Add Serialize and Scalar to Serial, update serial array visibility, and fix serialize scalar.
